Add files via upload
In this code, we generate two ER networks using the G(n,m) model with n=50 and m=2 (i.e., each node has 2 edges). We randomly select 10 nodes from the first network as "interactors" and select the same nodes from the second network as interactors. We then merge the two networks by creating edges between the interactors in the two networks and between the interactors and some random non-interactors.