Add files via upload
In a merged network, we have two ER networks that are connected by a set of "interactor" nodes that have edges to nodes in both ER networks. This structure can potentially create additional paths for the flow of information or other processes, which may affect the behavior of the network under bond percolation.